chore: move utilities and helpers out of deploy to be accessible by tasks
