+++ apps/smart-contracts/core/deploy/PrePOMarketFactory.ts
@@ -2,8 +2,8 @@
 import { DeployFunction } from 'hardhat-deploy/types'
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 import { ContractAddressOrInstance } from '@openzeppelin/hardhat-upgrades/dist/utils'
-import { assertIsTestnetChain, recordDeployment } from './utils'
-import { fetchExistingCollateral, sendTxAndWait } from './helpers'
+import { assertIsTestnetChain, recordDeployment } from '../utils'
+import { fetchExistingCollateral, sendTxAndWait } from '../helpers'
 import { PrePOMarketFactory } from '../typechain'
 
 const deployFunction: DeployFunction = async function ({
