+++ apps/smart-contracts/core/deploy/SingleStrategyController.ts
@@ -1,7 +1,7 @@
 // eslint-disable no-console
 import { DeployFunction } from 'hardhat-deploy/types'
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
-import { assertIsTestnetChain } from './utils'
+import { assertIsTestnetChain } from '../utils'
 
 const deployFunction: DeployFunction = async function ({
   deployments,
