+++ apps/smart-contracts/core/utils/index.ts
